+++ public/assets/JS/studio/studio.js
@@ -1,9 +1,11 @@
 import { isLoggedIn, readUserData, userData,userId, waitForUser } from "../authentication.js";
 import { getFirestore, collection, query, where, orderBy, limit, getDocs,getDoc,doc } from "https://www.gstatic.com/firebasejs/12.2.1/firebase-firestore.js";
 import { showChannelPosts } from "./studioShowPosts.js";
-import {userManagementPage} from "./userManagementPage.js"
+import { userManagementPage } from "./userManagementPage.js"
+import { studioNewPostPage } from "./studioNewPostPage.js";
 import app from "../../../firebase.js";
 import { loadSidebar, sidebarMap,constructSidebarButtons, updateSidebarButtonStatus } from "../sidebar.js";
+import { constructNavigationMap, contextualBottomNavigation } from "../navigation.js";
 
 const urlParams = new URLSearchParams(window.location.search);
 const currentPage = urlParams.get("currentPage");
@@ -72,12 +74,18 @@ userData.ownedChannels.forEach(channel => {
                     <br>
 
                     <br>
-                    <div class = "lios-action-button studio-action-view-channel-posts"><span>View Posts</span></div>
+                    <div class = "lios-button-group">
+                        <div class = "lios-action-button studio-action-view-channel-posts"><span>View Posts</span></div>
+                        <div class = "lios-action-button studio-action-new-channel-post"><span>New Post</span></div>
+                    </div>
             </div>
         `;
         channels.appendChild(displayChannel);
         displayChannel.querySelector(".studio-action-view-channel-posts").addEventListener("click", () => {
             showChannelPosts(channelInfo.uid);
+        });
+        displayChannel.querySelector(".studio-action-new-channel-post").addEventListener("click", () => {
+            studioNewPostPage(channelInfo.uid);
         });
     })();
 });
@@ -94,9 +102,11 @@ const constructStudioSidebar = (async() => {
         {
             "label": "Dashboard",
             "icon": "<svg xmlns=\"http://www.w3.org/2000/svg\" width=\"24\" height=\"24\" viewBox=\"0 0 24 24\" fill=\"none\" stroke=\"currentColor\" stroke-width=\"2\" stroke-linecap=\"round\" stroke-linejoin=\"round\" class=\"lucide lucide-gauge-icon lucide-gauge\"><path d=\"m12 14 4-4\"/><path d=\"M3.34 19a10 10 0 1 1 17.32 0\"/></svg>",
-            "action": () => {
+            "action": async () => {
                 virtualDom.innerHTML = "";
                 virtualDom.appendChild(dashboard);
+                await constructNavigationMap();
+                contextualBottomNavigation();
             }
         }
         
@@ -106,16 +116,20 @@ const constructStudioSidebar = (async() => {
             {
                 "label": "Channels",
                 "icon": "<svg xmlns=\"http://www.w3.org/2000/svg\" width=\"24\" height=\"24\" viewBox=\"0 0 24 24\" fill=\"none\" stroke=\"currentColor\" stroke-width=\"2\" stroke-linecap=\"round\" stroke-linejoin=\"round\" class=\"lucide lucide-rss-icon lucide-rss\"><path d=\"M4 11a9 9 0 0 1 9 9\"/><path d=\"M4 4a16 16 0 0 1 16 16\"/><circle cx=\"5\" cy=\"19\" r=\"1\"/></svg>",
-                "action": () => {
+                "action": async () => {
                     virtualDom.innerHTML = "";
                     virtualDom.appendChild(channels);
+                    await constructNavigationMap();
+                    contextualBottomNavigation();
                 }
             },
             {
                 "label": "User Management",
                 "icon": "<svg xmlns=\"http://www.w3.org/2000/svg\" width=\"24\" height=\"24\" viewBox=\"0 0 24 24\" fill=\"none\" stroke=\"currentColor\" stroke-width=\"2\" stroke-linecap=\"round\" stroke-linejoin=\"round\" class=\"lucide lucide-user-cog-icon lucide-user-cog\"><path d=\"M10 15H6a4 4 0 0 0-4 4v2\"/><path d=\"m14.305 16.53.923-.382\"/><path d=\"m15.228 13.852-.923-.383\"/><path d=\"m16.852 12.228-.383-.923\"/><path d=\"m16.852 17.772-.383.924\"/><path d=\"m19.148 12.228.383-.923\"/><path d=\"m19.53 18.696-.382-.924\"/><path d=\"m20.772 13.852.924-.383\"/><path d=\"m20.772 16.148.924.383\"/><circle cx=\"18\" cy=\"15\" r=\"3\"/><circle cx=\"9\" cy=\"7\" r=\"4\"/></svg>",
-                "action": () => {
-                    userManagementPage()
+                "action": async() => {
+                    userManagementPage();
+                    await constructNavigationMap();
+                    contextualBottomNavigation();
                 }
             }
         );